Added ignoring mechanism, moved IGenericDataProvider to the bootstrap.
