+++ inspectit-oce-bootstrap/src/main/java/rocks/inspectit/oce/bootstrap/instrumentation/DoNotInstrumentMarker.java
@@ -1,0 +1,7 @@
+package rocks.inspectit.oce.bootstrap.instrumentation;
+
+/**
+ * Simple marker interface which is used to exclude generated / injected classes from instrumentation.
+ */
+public interface DoNotInstrumentMarker {
+}
