+++ inspectit-oce-bootstrap/src/main/java/rocks/inspectit/oce/bootstrap/instrumentation/IGenericDataProvider.java
@@ -1,8 +1,8 @@
-package rocks.inspectit.oce.core.instrumentation.dataprovider.generic;
+package rocks.inspectit.oce.bootstrap.instrumentation;
 
 /**
- * Interface implemented by all Data-Providers which are generated by the {@link DataProviderGenerator}
- * The generated class use the template {@link GenericDataProviderTemplate}.
+ * Interface implemented by all Data-Providers which are generated by the DataProviderGenerator.
+ * The generated class use the template GenericDataProviderTemplate.
  */
 public interface IGenericDataProvider {
