+++ inspectit-oce-core/src/test/java/rocks/inspectit/oce/core/instrumentation/dataprovider/DataProviderGeneratorIntTest.java
@@ -5,11 +5,11 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.annotation.DirtiesContext;
+import rocks.inspectit.oce.bootstrap.instrumentation.IGenericDataProvider;
 import rocks.inspectit.oce.core.SpringTestBase;
 import rocks.inspectit.oce.core.instrumentation.config.model.ResolvedGenericDataProviderConfig;
 import rocks.inspectit.oce.core.instrumentation.dataprovider.generic.DataProviderGenerator;
 import rocks.inspectit.oce.core.instrumentation.dataprovider.generic.GenericDataProviderTemplate;
-import rocks.inspectit.oce.core.instrumentation.dataprovider.generic.IGenericDataProvider;
 import rocks.inspectit.oce.core.instrumentation.injection.InjectedClass;
 import rocks.inspectit.oce.core.testutils.DummyClassLoader;
 import rocks.inspectit.oce.core.testutils.GcUtils;
